- Introduced new blade files for displaying and editing bank transfers.
- Updated existing views to reference third-party names correctly.
- Enhanced validation rules in requests to improve data integrity.
- Added new models and relationships for compliance tracking.
- Refactored existing code for better readability and maintainability.

+++ app/Http/Controllers/Insurance/BancassuranceReferralController.php
@@ -19,36 +19,35 @@ use Illuminate\Http\Request;
 class BancassuranceReferralController extends Controller
 {
 
-    public function index()
-    {
-        $this->authorize(PermissionEnum::BancassuranceReferralView, BancAssuranceReferral::class);
-        $referrals = BancAssuranceReferral::with(['insuranceProduct', 'preferredInsurer', 'assignedToUser',])
-            ->orderByDesc('Id')->get();
+public function index()
+{
+    $this->authorize(PermissionEnum::BancassuranceReferralView, BancAssuranceReferral::class);
+    $referrals = BancAssuranceReferral::with(['insuranceProduct','preferredInsurer','assignedToUser',])
+    ->orderByDesc('Id')->get();
 
-        return view('bancassurance.referrals.index', compact('referrals'));
-    }
-
-    public function create()
-    {
-        //$this->authorize(PermissionEnum::BancassuranceReferralView, BancAssuranceReferral::class);
-        $insurers = InsuranceProvider::all();
-        $users = User::with('employee')->get();
-        return view('bancassurance.referrals.create', compact('users', 'insurers'));
-    }
-
-
-    public function getProductsByInsurer($insurerId)
-    {
-        $products = InsuranceProduct::where('InsuranceProviderID', $insurerId)->get();
-        return response()->json($products);
-    }
+    return view('bancassurance.referrals.index', compact('referrals'));
+}
+public function create()
+{
+    $this->authorize(PermissionEnum::BancassuranceReferralCreate, BancAssuranceReferral::class);
+    $insurers = InsuranceProvider::all();
+    $users = User::with('employee')->get();
+    return view('bancassurance.referrals.create', compact('users', 'insurers'));
+}
 
 
-    public function store(BancAssuranceReferralRequest $request)
-    {
-//$this->authorize(PermissionEnum::BancassuranceReferralCreate, BancAssuranceReferral::class);
-        $user = auth()->user();
-        $validated = $request->validated();
+public function getProductsByInsurer($insurerId)
+{
+    $products = InsuranceProduct::where('InsuranceProviderID', $insurerId)->get();
+    return response()->json($products);
+}
+
+
+public function store(BancAssuranceReferralRequest $request)
+{
+    $this->authorize(PermissionEnum::BancassuranceReferralCreate, BancAssuranceReferral::class);
+    $user = auth()->user();
+    $validated = $request->validated();
 
         // Handle nullable relationships safely
         $ReferredBy = !empty($validated['ReferredBy']) ? User::findOrFail($validated['ReferredBy']) : $user;
@@ -86,22 +85,22 @@ class BancassuranceReferralController extends Controller
             ->with('success', 'Referral submitted!');
     }
 
-    public function edit($Id)
-    {
-        $this->authorize(PermissionEnum::BancassuranceReferralUpdate, BancAssuranceReferral::class);
-        $referral = BancAssuranceReferral::with(['insuranceProduct', 'preferredInsurer', 'assignedToUser'])->findOrFail($Id);
-        $insuranceproducts = InsuranceProduct::all();
-        $insurers = InsuranceProvider::all();
-        $users = User::with('employee')->get();
+public function edit($Id)
+{
+    $this->authorize(PermissionEnum::BancassuranceReferralUpdate, BancAssuranceReferral::class);
+    $referral = BancAssuranceReferral::with(['insuranceProduct', 'preferredInsurer', 'assignedToUser'])->findOrFail($Id);
+    $insuranceproducts = InsuranceProduct::all();
+    $insurers = InsuranceProvider::all();
+    $users = User::with('employee')->get();
 
         return view('bancassurance.referrals.edit', compact('referral', 'users', 'insurers', 'insuranceproducts'));
     }
 
-    public function update(BancAssuranceReferralRequest $request, $Id)
-    {
-        $this->authorize(PermissionEnum::BancassuranceReferralUpdate, BancAssuranceReferral::class);
-        $user = auth()->user();
-        $validated = $request->validated();
+public function update(BancAssuranceReferralRequest $request, $Id)
+{  
+    $this->authorize(PermissionEnum::BancassuranceReferralUpdate, BancAssuranceReferral::class);
+    $user = auth()->user();
+    $validated = $request->validated();
 
         $referral = BancAssuranceReferral::where('Id', $Id)->firstOrFail();
 
@@ -141,20 +140,20 @@ class BancassuranceReferralController extends Controller
             ->with('success', 'Referral updated successfully!');
     }
 
-    public function show($Id)
-    {
-        $this->authorize(PermissionEnum::BancassuranceReferralView, BancAssuranceReferral::class);
-        $referral = BancAssuranceReferral::with('employee')->findOrFail($Id);
-        return view('bancassurance.referrals.show', compact('referral'));
-    }
+public function show($Id)
+{
+    $this->authorize(PermissionEnum::BancassuranceReferralView, BancAssuranceReferral::class);
+    $referral = BancAssuranceReferral::with('employee')->findOrFail($Id);
+    return view('bancassurance.referrals.show', compact('referral'));
+}
 
-    public function assignList()
-    {
-        $this->authorize(PermissionEnum::BancassuranceReferralView, BancAssuranceReferral::class);
-        $referrals = BancAssuranceReferral::with('insuranceProduct')
-            ->whereNull('AssignedTo')
-            ->where('Status', InsuranceReferralStatus::Pending->value) // Assuming 'P' stands for 'Pending'
-            ->orderByDesc('Id')->get();
+public function assignList()
+{
+    $this->authorize(PermissionEnum::BancassuranceReferralView, BancAssuranceReferral::class);
+    $referrals = BancAssuranceReferral::with('insuranceProduct')
+        ->whereNull('AssignedTo')
+        ->where('Status',InsuranceReferralStatus::Pending->value) // Assuming 'P' stands for 'Pending'
+        ->orderByDesc('Id')->get();
 
         $employees = Employee::whereNull('DeletedOn')
             ->select('Id', 'FirstName', 'LastName')
@@ -164,12 +163,12 @@ class BancassuranceReferralController extends Controller
     }
 
 
-    public function assign(Request $request, $Id)
-    {
-        $this->authorize(PermissionEnum::BancassuranceReferralCreate, BancAssuranceReferral::class);
-        $request->validate([
-            'AssignedTo' => 'required|exists:t_Employees,Id',
-        ]);
+public function assign(Request $request, $Id)
+{
+    $this->authorize(PermissionEnum::BancassuranceReferralCreate, BancAssuranceReferral::class);
+    $request->validate([
+        'AssignedTo' => 'required|exists:t_Employees,Id',
+    ]);
 
         $referral = BancAssuranceReferral::findOrFail($Id);
 
@@ -183,18 +182,20 @@ class BancassuranceReferralController extends Controller
         return redirect()->back()->with('success', 'Referral assigned successfully.');
     }
 
-    public function performanceView()
-    {
-        $this->authorize(PermissionEnum::BancassuranceReferralView, BancAssuranceReferral::class);
-        $referrals = BancAssuranceReferral::with(['referredByEmployee.branch'])->get();
+public function performanceView()
+{
+    $this->authorize(PermissionEnum::BancassuranceReferralView, BancAssuranceReferral::class);
+    $referrals = BancAssuranceReferral::with(['referredByEmployee.employee.branch'])->get();
 
         $grouped = $referrals->groupBy(function ($referral) {
             return $referral->referredByEmployee?->Id ?? 'Unknown';
         });
 
         $performance = $grouped->map(function ($items) {
-            $employee = $items->first()->referredByEmployee;
+            $employee = $items->first()->referredByEmployee->employee;
             $branch = $employee?->branch;
+
+           // dd($employee, $branch);
 
             return [
                 'StaffName' => $employee ? $employee->FirstName . ' ' . $employee->LastName : 'Unknown',
@@ -208,13 +209,18 @@ class BancassuranceReferralController extends Controller
         return view('bancassurance.referrals.performance', ['performance' => $performance]);
     }
 
-    public function destroy($Id)
-    {
-        $this->authorize(PermissionEnum::BancassuranceReferralDelete, BancAssuranceReferral::class);
-        $referral = BancAssuranceReferral::findOrFail($Id);
-        $referral->DeletedBy = Auth()->Id();
-        $referral->save();
-        $referral->delete();
+Public function destroy($Id)
+{
+    $this->authorize(PermissionEnum::BancassuranceReferralDelete, BancAssuranceReferral::class);
+    $referral = BancAssuranceReferral::findOrFail($Id);
+
+    if ($referral->customerreferral()->exists()) {
+        return redirect()->back()
+        ->withErrors(['error' => 'This Property Block is in use and cannot be deleted.']);
+    }  
+    $referral->DeletedBy = Auth()->Id();
+    $referral->save();
+    $referral->delete();
 
         activity()->causedBy(auth()->user()->Id)->performedOn($referral)
             ->event('delete')->log("Deleted Bank Assurance Referral {$referral->Id}.");
