+++ app/Http/Controllers/Property/PropertyAttachmentsController.php
@@ -18,13 +18,14 @@ class PropertyAttachmentsController extends Controller
     //
     public function index()
     {
+        $this->authorize(PermissionEnum::PropertyAttachmentsView, PropertyAttachments::class);
         $propertyattachments = PropertyAttachments::all();
 
         return view('property.propertyregistry.propertyattachments.index', compact('propertyattachments'));
     }
 
     public function create(){
-        //$this->authorize(PermissionEnum::PropertyAttachmentsCreate, PropertyAttachments::class);
+        $this->authorize(PermissionEnum::PropertyAttachmentsCreate, PropertyAttachments::class);
         $properties = PropertyRegistry::all();
         $documenttypes = CodeDetail::where('CodeID', 'DocumentType')->get();
         return view('property.propertyregistry.propertyattachments.create', compact('properties', 'documenttypes'));
@@ -32,10 +33,9 @@ class PropertyAttachmentsController extends Controller
 
     public function store(PropertyAttachmentsRequest $request)
     {
-        //$this->authorize(PermissionEnum::PropertyAttachmentsCreate, PropertyAttachments::class);
-        //dd($request->all());
+        $this->authorize(PermissionEnum::PropertyAttachmentsCreate, PropertyAttachments::class);
         $validated = $request->validated();
-
+        
         $PropertyID = PropertyRegistry::findOrFail($validated['PropertyID']);
         $DocumentType = CodeDetail::findOrFail($validated['DocumentType']);
         $user = Auth::user();
@@ -49,13 +49,14 @@ class PropertyAttachmentsController extends Controller
             $request->user(),
             $uploadedFile
         );
-        }
+    }
         return redirect()->route('attachments.index')->with('success', 'Property attachment created successfully');
     }
 
     public function edit($Id)
     {
-        // $this->authorize(PermissionEnum::PropertyAttachmentsView, PropertyAttachments::class);
+
+    $this->authorize(PermissionEnum::PropertyAttachmentsView, PropertyAttachments::class);
     $propertyattachments = PropertyAttachments::findOrFail($Id);
     $properties = PropertyRegistry::all();
     $documenttypes = CodeDetail::where('CodeID', 'DocumentType')->get();
@@ -64,15 +65,17 @@ class PropertyAttachmentsController extends Controller
     return view('property.propertyregistry.propertyattachments.edit', compact('propertyattachments', 'properties','documenttypes'));
     }
 
-    public function update(PropertyAttachmentsRequest $request, $id)
-    {
-        $validated = $request->validated();
+public function update(PropertyAttachmentsRequest $request, $id)
+{
+    $this->authorize(PermissionEnum::PropertyAttachmentsUpdate, PropertyAttachments::class);
 
-        // Fetch model instances
-        $PropertyID = PropertyRegistry::findOrFail($validated['PropertyID']);
-        $DocumentType = CodeDetail::findOrFail($validated['DocumentType']);
-        $user = Auth::user();
+    $validated = $request->validated();
 
+    // Fetch model instances
+    $PropertyID   = PropertyRegistry::findOrFail($validated['PropertyID']);
+    $DocumentType = CodeDetail::findOrFail($validated['DocumentType']);
+    $user         = Auth::user();
+    
 
     DB::beginTransaction();
 
@@ -120,7 +123,7 @@ class PropertyAttachmentsController extends Controller
 
 public function destroy($Id)
 {
-    //    $this->authorize(PermissionEnum::PropertyAttachmentsDelete, PropertyAttachmentsRequest::class);
+    $this->authorize(PermissionEnum::PropertyAttachmentsDelete, PropertyAttachmentsRequest::class);
     try {
         $propertyattachments = PropertyAttachments::findOrFail($Id);
         $propertyattachments->delete();
